config: switch default mixed_dtype from float16 to bfloat16
bfloat16 has the same exponent range as float32 so no loss scaling is
needed (GradScaler is already gated on float16 in training.py).
Avoids overflow/underflow issues on A100s.

+++ lom/config.py
@@ -70,7 +70,7 @@ class TrainCfg:
     eval_iters: int = 50
     log_interval: int = 50
     compile_model: bool = True
-    mixed_dtype: str = "float16"
+    mixed_dtype: str = "bfloat16"
     seed: int = 42
     ckpt_dir: str = "checkpoints"
     resume: bool = False
